requirements.txt
# ============================================================
# requirements.txt — Python dependencies
# ============================================================
#
# Versions are pinned with ~= (compatible release):
# ~= 0.110.0  means  >= 0.110.0, < 0.111.0
# This ensures reproducibility without blocking patch releases.
#
# To upgrade everything: pip install --upgrade -r requirements.txt
# To generate a precise lockfile: pip freeze > requirements.lock

# --- Web framework ---
fastapi~=0.111.0
# ASGI server (runs FastAPI)
uvicorn[standard]~=0.30.0

# --- Data validation ---
# Pydantic v2 is bundled with FastAPI, but we pin it explicitly
# to keep control over the version
pydantic~=2.7.0
pydantic-settings~=2.3.0

# --- Database ---
sqlalchemy[asyncio]~=2.0.30
# Async driver for PostgreSQL (used by the app at runtime)
asyncpg~=0.29.0
# Sync driver for PostgreSQL (used by Alembic to run migrations)
psycopg2-binary~=2.9.9
# Database migrations
alembic~=1.13.0

# --- Task queue (async federation) ---
celery[redis]~=5.4.0

# --- Async HTTP client (for sending AP activities to other servers) ---
httpx~=0.27.0

# --- Cryptography (HTTP Signatures for ActivityPub) ---
cryptography~=42.0.0

# --- Auth ---
# Password hashing
bcrypt~=4.1.0
# JWT
python-jose[cryptography]~=3.3.0

# --- Utilities ---
# Parsing of ISO 8601 durations (PT20M, etc.)
isodate~=0.6.1
# URL-friendly slugs from unicode text
python-slugify~=8.0.0

# --- Async file I/O ---
aiofiles>=23.0.0

# --- Testing ---
# Test framework
pytest~=8.2.0
# Async test support for pytest
pytest-asyncio~=0.23.0
# Async test utilities
anyio~=4.4.0

# --- Mail validation ---
fastapi-mail~=1.4.0
